Split FieldBackedImplementationInstaller into several smaller classes
